requirements.txt
flask
requests
pandas
mysql-connector-python
